+++ src/App.jsx
@@ -1,18 +1,16 @@
 
 import './App.css'
+import TodoList from './TodoList.jsx'
+import TodoForm from './TodoForm.jsx'
+
 
 function App() {
-  const todos = [
-    {id:1, title: "Clean Room"},
-    {id:2, title: "Do Homework"},
-    {id:3, title: "Wash Hair"}
-  ]
+
   return (
     <div>
       <h1>My Todos</h1>
-      <ul>
-        {todos.map(todo => <li key={todo.id}>{todo.title}</li>)}
-      </ul>
+      <TodoForm/>
+      <TodoList/>
     </div>
   )
 }
